@@ -13,12 +13,12 @@
   },
   "private": true,
   "dependencies": {
-    "@angular/common": "~13.0.0",
-    "@angular/core": "~13.0.0",
-    "@angular/forms": "~13.0.0",
-    "@angular/platform-browser": "~13.0.0",
-    "@angular/platform-browser-dynamic": "~13.0.0",
-    "@angular/router": "~13.0.0",
+    "@angular/common": "~14.3.0",
+    "@angular/core": "~14.3.0",
+    "@angular/forms": "~14.3.0",
+    "@angular/platform-browser": "~14.3.0",
+    "@angular/platform-browser-dynamic": "~14.3.0",
+    "@angular/router": "~14.3.0",
     "@awesome-cordova-plugins/geolocation": "^5.37.3",
     "@awesome-cordova-plugins/native-geocoder": "^5.37.3",
     "@capacitor/android": "3.3.4",
@@ -41,9 +41,9 @@
     "@angular-eslint/eslint-plugin-template": "~13.0.1",
     "@angular-eslint/template-parser": "~13.0.1",
     "@angular/cli": "^17.0.5",
-    "@angular/compiler": "~13.0.0",
-    "@angular/compiler-cli": "~13.0.0",
-    "@angular/language-service": "~13.0.0",
+    "@angular/compiler": "~14.3.0",
+    "@angular/compiler-cli": "~14.3.0",
+    "@angular/language-service": "~14.3.0",
     "@capacitor/cli": "^5.5.1",
     "@ionic/angular-toolkit": "^5.0.0",
     "@types/jasmine": "~3.6.0",
@@ -65,7 +65,7 @@
     "karma-jasmine-html-reporter": "^1.5.0",
     "protractor": "^3.3.0",
     "ts-node": "~8.3.0",
-    "typescript": "~4.4.4"
+    "typescript": "~4.8.4"
   },
   "description": "An Ionic project"
 }
